fix(atlas): move atlas_db out of scripts/lexicon so it doesn't churn the manifest fingerprint
The manifest-fingerprint hook hashes scripts/lexicon/*.py and treats any change
as 'manifest may be stale', which failed the frontend hydrate check (pointer
fingerprint mismatch) — atlas_db.py builds a SEPARATE DB and does not generate
the JSON manifest. Move it to scripts/atlas/ (same pattern as the regen script
living outside scripts/lexicon/), revert the fingerprint churn. PR now touches
no site/ file, so the frontend workflow correctly skips.
